+++ example/prometheus/main.go
@@ -29,8 +29,11 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/exporters/prometheus"
 	api "go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric"
 )
+
+const meterName = "github.com/open-telemetry/opentelemetry-go/example/prometheus"
 
 func main() {
 	rng := rand.New(rand.NewSource(time.Now().UnixNano()))
@@ -43,8 +46,23 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-	provider := metric.NewMeterProvider(metric.WithReader(exporter))
-	meter := provider.Meter("github.com/open-telemetry/opentelemetry-go/example/prometheus")
+	provider := metric.NewMeterProvider(
+		metric.WithReader(exporter),
+		// View to customize histogram buckets and rename a single histogram instrument.
+		metric.WithView(metric.NewView(
+			metric.Instrument{
+				Name:  "baz",
+				Scope: instrumentation.Scope{Name: meterName},
+			},
+			metric.Stream{
+				Name: "new_baz",
+				Aggregation: metric.AggregationExplicitBucketHistogram{
+					Boundaries: []float64{64, 128, 256, 512, 1024, 2048, 4096},
+				},
+			},
+		)),
+	)
+	meter := provider.Meter(meterName)
 
 	// Start the prometheus HTTP server and pass the exporter Collector to it
 	go serveMetrics()
@@ -75,14 +93,14 @@ func main() {
 	}
 
 	// This is the equivalent of prometheus.NewHistogramVec
-	histogram, err := meter.Float64Histogram("baz", api.WithDescription("a very nice histogram"))
+	histogram, err := meter.Float64Histogram("baz", api.WithDescription("a histogram with custom buckets and rename"))
 	if err != nil {
 		log.Fatal(err)
 	}
-	histogram.Record(ctx, 23, opt)
-	histogram.Record(ctx, 7, opt)
-	histogram.Record(ctx, 101, opt)
-	histogram.Record(ctx, 105, opt)
+	histogram.Record(ctx, 136, opt)
+	histogram.Record(ctx, 64, opt)
+	histogram.Record(ctx, 701, opt)
+	histogram.Record(ctx, 830, opt)
 
 	ctx, _ = signal.NotifyContext(ctx, os.Interrupt)
 	<-ctx.Done()